+++ src/ewoksserver/app/routes/tasks/__init__.py
@@ -4,4 +4,5 @@ routers = {
     (1, 0, 0): _router,
     (1, 1, 0): _router,
     (2, 0, 0): _router,
+    (2, 1, 0): _router,
 }